without orders

+++ src/Controller/Admin/AvailabilityController.php
@@ -5,7 +5,6 @@ namespace App\Controller\Admin;
 use App\Entity\Availability;
 use App\Form\Admin\Availability\AvailabilityType;
 use App\Repository\AvailabilityRepository;
-use App\Services\Common\TranslationRecipient;
 use Doctrine\ORM\EntityManagerInterface;
 use Gedmo\Translatable\Entity\Translation;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -13,6 +12,7 @@ use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
@@ -25,38 +25,34 @@ class AvailabilityController extends AbstractController
     /**
      * @Route("", name="_all")
      * @param AvailabilityRepository $availabilityRepository
-     * @param TranslationRecipient $translationRecipient
      * @return Response
      */
-    public function availabilityAll(AvailabilityRepository $availabilityRepository, TranslationRecipient $translationRecipient)
+    public function availabilityAll(AvailabilityRepository $availabilityRepository)
     {
-        $availabilities = [];
-        $availabilitiesAll = $availabilityRepository->findAll();
-        foreach ($availabilitiesAll as $availability) {
-            $availabilities[] = $translationRecipient->getTranslatedEntity($availability);
-        }
+        $user = $this->getUser();
 
         return $this->render('admin/availability/all.html.twig', [
             'controller_name' => 'AvailabilityController',
-            'availabilities' => $availabilities,
+            'availabilities' => $availabilityRepository->adminAvailabilitiesList($user),
         ]);
     }
 
     /**
      * @Route("/{id}", name="_single", requirements={"id"="\d+"})
      * @param Availability $availability
-     * @param TranslationRecipient $translationRecipient
+     * @param TranslatorInterface $translator
      * @return Response
      */
-    public function availabilitySingle(Availability $availability, TranslationRecipient $translationRecipient)
+    public function availabilitySingle(Availability $availability, TranslatorInterface $translator)
     {
-        $translation = $translationRecipient->getTranslation($availability);
-
-        return $this->render('admin/availability/single.html.twig', [
-            'controller_name' => 'AvailabilityController',
-            'availability' => $availability,
-            'translation' => $translation,
-        ]);
+        if ($this->getUser() === $availability->getUser() || $this->isGranted('ROLE_SUPERADMIN')) {
+            return $this->render('admin/availability/single.html.twig', [
+                'controller_name' => 'AvailabilityController',
+                'availability' => $availability,
+            ]);
+        } else {
+            throw new AccessDeniedException($translator->trans('У вас нет доступа для данной операции'));
+        }
     }
 
     /**
@@ -87,6 +83,9 @@ class AvailabilityController extends AbstractController
                 ->translate($availability, 'short_description', 'uk', $arrData['translation_short_description'])
                 ->translate($availability, 'description', 'uk', $arrData['translation_description']);
 
+            $user = $this->getUser();
+            $availability->setUser($user);
+
             $entityManager->persist($availability);
             $entityManager->flush();
 
@@ -113,34 +112,40 @@ class AvailabilityController extends AbstractController
      */
     public function availabilityEdit(Availability $availability, Request $request, EntityManagerInterface $entityManager, TranslatorInterface $translator)
     {
-        $form = $this->createForm(AvailabilityType::class, $availability, [
-            'action' => $this->generateUrl('admin_availability_edit', ['id' => $availability->getId()]),
-            'method' => 'post',
-        ]);
+        $user = $this->getUser();
 
-        $form->handleRequest($request);
+        if ($user === $availability->getUser() || $this->isGranted('ROLE_SUPERADMIN')) {
+            $form = $this->createForm(AvailabilityType::class, $availability, [
+                'action' => $this->generateUrl('admin_availability_edit', ['id' => $availability->getId()]),
+                'method' => 'post',
+            ]);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $arrData = $request->request->get('availability');
-            $repoTranslation = $entityManager->getRepository(Translation::class);
-            $repoTranslation->translate($availability, 'name', 'uk', $arrData['translation_name'])
-                ->translate($availability, 'short_description', 'uk', $arrData['translation_short_description'])
-                ->translate($availability, 'description', 'uk', $arrData['translation_description']);
+            $form->handleRequest($request);
 
-            $entityManager->persist($availability);
-            $entityManager->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $arrData = $request->request->get('availability');
+                $repoTranslation = $entityManager->getRepository(Translation::class);
+                $repoTranslation->translate($availability, 'name', 'uk', $arrData['translation_name'])
+                    ->translate($availability, 'short_description', 'uk', $arrData['translation_short_description'])
+                    ->translate($availability, 'description', 'uk', $arrData['translation_description']);
 
-            $message = $translator->trans('Доступность успешно изменена');
-            $this->addFlash('success', $message);
+                $entityManager->persist($availability);
+                $entityManager->flush();
 
-            return $this->redirectToRoute('admin_availability_single', ['id' => $availability->getId()]);
+                $message = $translator->trans('Доступность успешно изменена');
+                $this->addFlash('success', $message);
+
+                return $this->redirectToRoute('admin_availability_single', ['id' => $availability->getId()]);
+            }
+
+            return $this->render('admin/availability/add.html.twig', [
+                'controller_name' => 'AvailabilityController',
+                'form_add' => $form->createView(),
+                'title' => $translator->trans('Редактирование доступности продукта'),
+            ]);
+        } else {
+            throw new AccessDeniedException($translator->trans('У вас нет доступа для данной операции'));
         }
-
-        return $this->render('admin/availability/add.html.twig', [
-            'controller_name' => 'AvailabilityController',
-            'form_add' => $form->createView(),
-            'title' => $translator->trans('Редактирование доступности продукта'),
-        ]);
     }
 
     /**
@@ -152,13 +157,19 @@ class AvailabilityController extends AbstractController
      */
     public function availabilityDelete(Availability $availability, EntityManagerInterface $entityManager, TranslatorInterface $translator)
     {
-        $entityManager->remove($availability);
-        $entityManager->flush();
+        $user = $this->getUser();
 
-        $message = $translator->trans('Доступность успешно удалена');
+        if ($user === $availability->getUser() || $this->isGranted('ROLE_SUPERADMIN')) {
+            $entityManager->remove($availability);
+            $entityManager->flush();
 
-        $this->addFlash('success', $message);
+            $message = $translator->trans('Доступность успешно удалена');
 
-        return $this->redirectToRoute('admin_availability_all');
+            $this->addFlash('success', $message);
+
+            return $this->redirectToRoute('admin_availability_all');
+        } else {
+            throw new AccessDeniedException($translator->trans('У вас нет доступа для данной операции'));
+        }
     }
 }
